src/main/webapp/app/entities/admin-cron/route/admin-cron-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IAdminCron, AdminCron } from '../admin-cron.model';
import { AdminCronService } from '../service/admin-cron.service';

import { AdminCronRoutingResolveService } from './admin-cron-routing-resolve.service';

describe('AdminCron routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: AdminCronRoutingResolveService;
  let service: AdminCronService;
  let resultAdminCron: IAdminCron | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(AdminCronRoutingResolveService);
    service = TestBed.inject(AdminCronService);
    resultAdminCron = undefined;
  });

  describe('resolve', () => {
    it('should return IAdminCron returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultAdminCron = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultAdminCron).toEqual({ id: 123 });
    });

    it('should return new IAdminCron if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultAdminCron = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultAdminCron).toEqual(new AdminCron());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as AdminCron })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultAdminCron = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultAdminCron).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
